+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## [0.10.0] - 2020-03-18
+### Added
+- `Document::input_text`
+
 ### Changed
 - `Ancestors`, `PrevSiblings`, `NextSiblings`, `FirstChildren` and `LastChildren`
   were replaced with `AxisIter`.
-  
+
 ### Fixed
 - Root node range.
 
@@ -121,7 +125,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `ExpandedName::has_namespace`. `ExpandedName::namespace` will return `Option` now.
 - `Error::NestedEntityReference`.
 
-[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.9.1...HEAD
+[Unreleased]: https://github.com/RazrFalcon/roxmltree/compare/v0.10.0..HEAD
+[0.10.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.9.1...v0.10.0
 [0.9.1]: https://github.com/RazrFalcon/roxmltree/compare/v0.9.0...v0.9.1
 [0.9.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/RazrFalcon/roxmltree/compare/v0.7.3...v0.8.0